// src/entities/achievement/types.ts
import { Timestamp } from "firebase/firestore";

export interface Badge {
  id: string;
  title: string;
  color?: string;
  icon?: string;
}

export interface Achievement {
  id: string;
  title: string;
  desc: string;
  date: Timestamp;
  cover: string;
  link?: string;
  badges?: Badge[];
}

export interface AchievementListResponse {
  data: Achievement[];
  meta: {
    pagination: {
      page: number;
      pageSize: number;
      pageCount: number;
      total: number;
    };
  };
}